backend/requirements.txt
flask
faiss-cpu
sentence-transformers
pypdf
numpy
